+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to [@camunda/linting](https://github.com/camunda/linting) ar
 
 ## Unreleased
 
+___Note:__ Yet to be released changes appear here._
+
+## 3.56.1
+
 * `FIX`: recommend `camunda.secrets.*` for legacy and deprecated secret expressions from Camunda 8.10 ([camunda/camunda-modeler#6156](https://github.com/camunda/camunda-modeler/issues/6156))
 * `DEPS`: update to `bpmnlint-plugin-camunda-compat@2.60.4`
-
-___Note:__ Yet to be released changes appear here._
 
 ## 3.56.0